+++ test/oracle/OracleMiddleware_GetPriceTest.t.sol
@@ -16,12 +16,12 @@ contract OracleMiddleware_GetPriceTest is OracleMiddleware_BaseTest {
   function testCorrectness_GetLatestPrice() external {
     // Should get price via PythAdapter successfully.
     // For more edge cases see PythAdapter_GetPriceTest.t.sol
-    (uint maxPrice, uint lastUpdate) = oracleMiddleware.getLatestPrice(
+    (uint256 maxPrice, uint256 lastUpdate) = oracleMiddleware.getLatestPrice(
       address(wbtc).toBytes32(),
       true,
       1 ether
     );
-    (uint minPrice, ) = oracleMiddleware.getLatestPrice(
+    (uint256 minPrice, ) = oracleMiddleware.getLatestPrice(
       address(wbtc).toBytes32(),
       false,
       1 ether
